Update patch set 3
Patch Set 3:

on cm13 times i noticed that changing the network mode from settings ended up in an infinite network loop (activating and deactivating all the time) if you go into *#*#4636#*#* and until you get out, that stopped to happen after changing the default property, so i just replicate this on 14.1 to be sure i have no network problems.
Anyway, peoples might not even know or look into the network mode settings and they will use the default CDMA/LTE etc.. forever, but that's not my purpose, i just pushed this to use it for me or whoever wants to use it, i have no intention of trying to get it merged.

Patch-set: 3